+++ log/src/Log/Backend/ElasticSearch.hs
@@ -1,0 +1,5 @@
+module Log.Backend.ElasticSearch
+    {-# DEPRECATED "Use directly Log.Backend.ElasticSearch.V1 or V5" #-}
+    ( module Log.Backend.ElasticSearch.V1 ) where
+
+import Log.Backend.ElasticSearch.V1