+++ log/src/Log/Backend/ElasticSearch/Internal.hs
@@ -1,0 +1,5 @@
+module Log.Backend.ElasticSearch.Internal
+    {-# DEPRECATED "Use directly Log.Backend.ElasticSearch.V1 or V5" #-}
+    ( module Log.Backend.ElasticSearch.V1.Internal ) where
+
+import Log.Backend.ElasticSearch.V1.Internal