+++ log/src/Log/Backend/ElasticSearch/Lens.hs
@@ -1,0 +1,5 @@
+module Log.Backend.ElasticSearch.Lens
+    {-# DEPRECATED "Use directly Log.Backend.ElasticSearch.V1 or V5" #-}
+    ( module Log.Backend.ElasticSearch.V1.Lens ) where
+
+import Log.Backend.ElasticSearch.V1.Lens