+++ log-elasticsearch/CHANGELOG.md
@@ -1,8 +1,11 @@
-# log-elasticsearch-0.8.1 (2016-03-27)
+# log-elasticsearch-0.9.0.0 (2017-05-04)
+* Now works with bloodhound-0.14.0.0 (#30).
+
+# log-elasticsearch-0.8.1 (2017-03-27)
 * Log.Backend.ElasticSearch.Internal now exports 'EsUsername' and
   'EsPassword'.
 
-# log-elasticsearch-0.8 (2016-03-16)
+# log-elasticsearch-0.8 (2017-03-16)
 * Made ElasticSearchConfig an abstract type (#27).
 * Added support for HTTPS and basic auth (#26).